+++ crates/models/dependent/src/core_systems.rs
@@ -71,7 +71,7 @@ pub struct CoreDetailsProviderSpecifics {
 
 #[derive(PartialEq, Eq, Clone, Debug, SimpleObject, Serialize, Deserialize)]
 pub struct CoreDetails {
-    pub page_size: i32,
+    pub page_size: u64,
     pub version: String,
     pub docs_link: String,
     pub oidc_enabled: bool,
