+++ crates/providers/itunes/src/lib.rs
@@ -161,7 +161,7 @@ impl MediaProvider for ITunesService {
 
     async fn metadata_search(
         &self,
-        page: i32,
+        page: u64,
         query: &str,
         _display_nsfw: bool,
         _source_specifics: &Option<MetadataSearchSourceSpecifics>,
