+++ crates/providers/tvdb/src/movies.rs
@@ -38,7 +38,7 @@ impl TvdbMovieService {
 impl MediaProvider for TvdbMovieService {
     async fn metadata_search(
         &self,
-        page: i32,
+        page: u64,
         query: &str,
         _display_nsfw: bool,
         _source_specifics: &Option<MetadataSearchSourceSpecifics>,
