+++ crates/providers/tvdb/src/shows.rs
@@ -38,7 +38,7 @@ impl TvdbShowService {
 impl MediaProvider for TvdbShowService {
     async fn metadata_search(
         &self,
-        page: i32,
+        page: u64,
         query: &str,
         _display_nsfw: bool,
         _source_specifics: &Option<MetadataSearchSourceSpecifics>,
