+++ crates/services/collection/src/recommendation_operations.rs
@@ -99,7 +99,7 @@ pub async fn collection_recommendations(
         items: paginator.fetch_page(page - 1).await?,
         details: SearchDetails {
             total_items: number_of_items.try_into().unwrap(),
-            next_page: (page < number_of_pages).then(|| (page + 1) as i32),
+            next_page: (page < number_of_pages).then(|| page + 1),
         },
     })
 }
